requirements.txt
pycryptodome==3.19.0
Pillow==10.1.0
numpy==1.24.3
